utils: rename util -> utils and switch to a module
