+++ src/crates.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::str::FromStr;
 use tar::Archive;
-use util;
+use utils;
 
 const CRATES_ROOT: &str = "https://crates-io.s3-us-west-1.amazonaws.com/crates";
 
@@ -146,7 +146,7 @@ pub fn prepare_crate(krate: &Crate) -> Result<()> {
                 repo.slug(),
                 dir.display()
             );
-            util::copy_dir(&repo.mirror_dir(), &dir)?;
+            utils::copy_dir(&repo.mirror_dir(), &dir)?;
         }
     }
 
@@ -173,7 +173,7 @@ fn dl_registry(name: &str, vers: &str, dir: &Path) -> Result<()> {
     let r = unpack_without_first_dir(&mut tar, dir).chain_err(|| "unable to unpack crate tarball");
 
     if r.is_err() {
-        let _ = util::remove_dir_all(dir);
+        let _ = utils::remove_dir_all(dir);
     }
 
     r
